requirements.txt
asgiref==3.8.1
Django==5.1
django-AutoDateTimeFields==0.9.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1